Cleaning up line drawing and fixing a few related bugs

+++ src/lib/sputter/render/rasterizer.h
@@ -3,7 +3,7 @@
 #include <cstdint>
 
 namespace sputter { namespace render {
-    void DrawVerticalLine(uint8_t x, uint8_t y0, uint8_t y1, uint8_t* pData, uint8_t stride);
-    void DrawHorizontalLine(uint8_t y, uint8_t x0, uint8_t x1, uint8_t* pData, uint8_t stride);
-    void DrawLine(uint8_t x0, uint8_t y0, uint8_t x1, uint8_t y1, uint8_t* pData, uint8_t stride);
+    void DrawVerticalLine(int16_t x, int16_t y0, int16_t y1, uint8_t* pData, int16_t stride);
+    void DrawHorizontalLine(int16_t y, int16_t x0, int16_t x1, uint8_t* pData, int16_t stride);
+    void DrawLine(int16_t x0, int16_t y0, int16_t x1, int16_t y1, uint8_t* pData, int16_t stride);
 }}
